refactor[#188] : util.ts 를 index.ts 로 바꿈
