+++ .prettierrc
@@ -11,8 +11,8 @@
 	"requirePragma": false,
 	"semi": true,
 	"singleQuote": true,
-	"tabWidth": 4,
+	"tabWidth": 2,
 	"trailingComma": "none",
-	"useTabs": true,
+	"useTabs": false,
 	"vueIndentScriptAndStyle": false
 }
